Change communication protocol with radio(xbee) to  UART again
To use SPI communication of xbee is necessary hold reset and set dout pin
to high until some other pin turn to high, then we should set dout to low and
release reset.

So is simpler use uart, as we only send small strings we will
not have performance issues.